requirements.txt
Flask==3.0.0
gunicorn==21.2.0
google-generativeai==0.5.2
